fix(tests): fix annotations to match default JUnit version
 - Remove JUnit4 annotation `@RunWith(SpringRunner.class)`.
 - Add JUnit5 annotation `@ActiveProfiles("test")` (add
`application-test.properties for ` test-specifi configuration).
