+++ Poseiden-skeleton/src/test/java/com/pcs/tradingapp/BidTests.java
@@ -1,0 +1,46 @@
+package com.pcs.tradingapp;
+
+import com.pcs.tradingapp.domain.BidList;
+import com.pcs.tradingapp.repositories.BidListRepository;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+
+import java.util.List;
+import java.util.Optional;
+
+@ActiveProfiles("test")
+@SpringBootTest
+public class BidTests {
+
+	@Autowired
+	private BidListRepository bidListRepository;
+
+	@Test
+	public void bidListTest() {
+		BidList bid = new BidList();
+
+		// Save
+		bid = bidListRepository.save(bid);
+		Assertions.assertNotNull(bid.getId());
+		Assertions.assertEquals(bid.getBidQuantity(), 10d, 10d);
+
+		// Update
+		bid.setBidQuantity(20d);
+		bid = bidListRepository.save(bid);
+		Assertions.assertEquals(bid.getBidQuantity(), 20d, 20d);
+
+		// Find
+		List<BidList> listResult = bidListRepository.findAll();
+		Assertions.assertTrue(listResult.size() > 0);
+
+		// Delete
+		Integer id = bid.getId();
+		bidListRepository.delete(bid);
+		Optional<BidList> bidList = bidListRepository.findById(id);
+		Assertions.assertFalse(bidList.isPresent());
+	}
+}
