+++ Poseiden-skeleton/src/test/java/com/pcs/tradingapp/CurvePointTests.java
@@ -1,0 +1,47 @@
+package com.pcs.tradingapp;
+
+import com.pcs.tradingapp.domain.CurvePoint;
+import com.pcs.tradingapp.repositories.CurvePointRepository;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+
+import java.util.List;
+import java.util.Optional;
+
+@ActiveProfiles("test")
+@SpringBootTest
+public class CurvePointTests {
+
+	@Autowired
+	private CurvePointRepository curvePointRepository;
+
+	@Test
+	public void curvePointTest() {
+		CurvePoint curvePoint = new CurvePoint();
+
+		// Save
+		curvePoint = curvePointRepository.save(curvePoint);
+		Assertions.assertNotNull(curvePoint.getId());
+		Assertions.assertTrue(curvePoint.getCurveId() == 10);
+
+		// Update
+		curvePoint.setCurveId(20);
+		curvePoint = curvePointRepository.save(curvePoint);
+		Assertions.assertTrue(curvePoint.getCurveId() == 20);
+
+		// Find
+		List<CurvePoint> listResult = curvePointRepository.findAll();
+		Assertions.assertTrue(listResult.size() > 0);
+
+		// Delete
+		Integer id = curvePoint.getId();
+		curvePointRepository.delete(curvePoint);
+		Optional<CurvePoint> curvePointList = curvePointRepository.findById(id);
+		Assertions.assertFalse(curvePointList.isPresent());
+	}
+
+}
