+++ Poseiden-skeleton/src/test/java/com/pcs/tradingapp/PasswordEncodeTest.java
@@ -1,0 +1,16 @@
+package com.pcs.tradingapp;
+
+import org.junit.jupiter.api.Test;
+
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+
+@SpringBootTest
+public class PasswordEncodeTest {
+    @Test
+    public void testPassword() {
+        BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
+        String pw = encoder.encode("123456");
+        System.out.println("[ "+ pw + " ]");
+    }
+}
