+++ Poseiden-skeleton/src/test/java/com/pcs/tradingapp/RuleTests.java
@@ -1,0 +1,46 @@
+package com.pcs.tradingapp;
+
+import com.pcs.tradingapp.domain.RuleName;
+import com.pcs.tradingapp.repositories.RuleNameRepository;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+
+import java.util.List;
+import java.util.Optional;
+
+@ActiveProfiles("test")
+@SpringBootTest
+public class RuleTests {
+
+	@Autowired
+	private RuleNameRepository ruleNameRepository;
+
+	@Test
+	public void ruleTest() {
+		RuleName rule = new RuleName();
+
+		// Save
+		rule = ruleNameRepository.save(rule);
+		Assertions.assertNotNull(rule.getId());
+		Assertions.assertTrue(rule.getName().equals("Rule Name"));
+
+		// Update
+		rule.setName("Rule Name Update");
+		rule = ruleNameRepository.save(rule);
+		Assertions.assertTrue(rule.getName().equals("Rule Name Update"));
+
+		// Find
+		List<RuleName> listResult = ruleNameRepository.findAll();
+		Assertions.assertTrue(listResult.size() > 0);
+
+		// Delete
+		Integer id = rule.getId();
+		ruleNameRepository.delete(rule);
+		Optional<RuleName> ruleList = ruleNameRepository.findById(id);
+		Assertions.assertFalse(ruleList.isPresent());
+	}
+}
