+++ Poseiden-skeleton/src/test/java/com/pcs/tradingapp/TradeTests.java
@@ -1,0 +1,46 @@
+package com.pcs.tradingapp;
+
+import com.pcs.tradingapp.domain.Trade;
+import com.pcs.tradingapp.repositories.TradeRepository;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+
+import java.util.List;
+import java.util.Optional;
+
+@ActiveProfiles("test")
+@SpringBootTest
+public class TradeTests {
+
+	@Autowired
+	private TradeRepository tradeRepository;
+
+	@Test
+	public void tradeTest() {
+		Trade trade = new Trade();
+
+		// Save
+		trade = tradeRepository.save(trade);
+		Assertions.assertNotNull(trade.getId());
+		Assertions.assertTrue(trade.getAccount().equals("Trade Account"));
+
+		// Update
+		trade.setAccount("Trade Account Update");
+		trade = tradeRepository.save(trade);
+		Assertions.assertTrue(trade.getAccount().equals("Trade Account Update"));
+
+		// Find
+		List<Trade> listResult = tradeRepository.findAll();
+		Assertions.assertTrue(listResult.size() > 0);
+
+		// Delete
+		Integer id = trade.getId();
+		tradeRepository.delete(trade);
+		Optional<Trade> tradeList = tradeRepository.findById(id);
+		Assertions.assertFalse(tradeList.isPresent());
+	}
+}
